only warn about artifact timestamp on attestation
Previously, a warning was issued whenever a kas build was executed and
the artifacts specified in the config file have not been created
in-between the start and the end of the build. However, these
artifacts can also be there from a previous build (or partial build),
hence the timestamps are expected to differ on local executions.

This patch moves this check to the attestation part, which is expected
to always start from a clean state (sstate is ok, but no lingering and
untracked files). There it is an important debugging mechanism.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>