do not always require artifacts to be generated
Previously, the build was marked failed in case at least one artifact
was not found after the bitbake returned. However, this is too strict,
as some build commands which are used during development (e.g. clean),
don't generate artifacts by design.

Instead, we now only require the artifacts to be present when also
generating build attestations. This patch updates the specification
(docs), the implementation as well as the test for missing artifacts.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>