ci: upload to current project registry
The container upload part of the github action is currently not
portable. By that, the CI on forks of kas fails as it obviously cannot
upload to the official kas container registry.

This patch replaces the hardcoded container registry path with the
predefined variable github.repository and by that makes the action
portable.

In theory, this also makes it possible to have a tracking rebuilder that
just mirrors the siemens/kas repository, rebuilds the containers using
the "official" workflow and compares the checksums.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>