auto-inject git credentials on gitlab ci
When running in the gitlab ci, we now automatically configure git to
clone the repositories on the CI server using https and the
CI_JOB_TOKEN. For backwards compatibility, this auto-injection is only
performed if no local git config or SSH config is provided.
This makes it easier for our users to use KAS in the Gitlab CI, as
things just work out of the box without any manual configuration needed.

For the Github CI this was already possible. By that, no change is
needed.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Tested-by: Frieder Schrempf <frieder.schrempf@kontron.de>
Reviewed-by: Frieder Schrempf <frieder.schrempf@kontron.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>