ci: redact non-reproducible annotations of container
Prior to this patch, only the layers itself are reproducible,
but not the manifests. By that, rebuilders (e.g. via github
action in forks) will generate manifests with a different digest.
The reason for that are non-reproducible annotations in the manifests.

This patch fixes this by redacting the problematic annotations.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>