use image tests as full-blown examples
The image tests already provide good and rather minimal examples. As we
currently lack fully working examples, we move the existing test code
into an examples dir and symlink these into the image-tests dir. Further,
we reference (include) them in the documentation.

Creating dedicated examples has been considered, but as these are likely
hard to maintain this idea was dropped.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>