introduce container registry login
Create a container registry authentication file that is
currently compatible with docker, podman and skopeo.
The format is described here [1].

When REGISTRY_AUTH_FILE is set then use that file and copy it into the
kas home directory.

If CI_REGISTRY, CI_REGISTRY_USER and CI_JOB_TOKEN is set then
the required login data is added to the login file.

[1] https://github.com/containers/image/blob/main/docs/containers-auth.json.5.md

Signed-off-by: Benedikt Niedermayr <benedikt.niedermayr@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>