kas-container image: re-add deterministic metadata
In 16e04461 we cleared the non-reproducible parts of the
org.opencontainers.image metadata that is added to the image manifest
(and index manifest). This however removed interesting information like
when the image was generated and where to find the sources from image.

This commit re-adds this information in a deterministic way. Instead of
created, we now use the stable committed timestamp (as a rebuild will be
bit-by-bit identical with that version anyways). For .source and .url we
use the official kas URL, so tracking rebuilders (like forks) still can
regenerate the image.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>