keep git committer identity if provided in .gitconfig
In case a .gitconfig file is provided by the user and this file contains
a user section, restore this after repo patching. By that, git operates
under the provided user in the shell command.

As this changes the semantics (git user is restored vs. git user is
overwritten), we also adapt the test.

Proposed-by: José Manuel Díez <jose.manuel.diez@gmail.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>