document difference between file and env credentials
Credentials that are provided via files always had to be explicitly
forwarded into the environment, while env var based ones were
automatically forwarded. This difference makes sense, as "hiding"
credential files is tricky for the user, while unsetting env vars is
easy.

To make this clear to our users, we add a note to the credential
handling documentation. While documenting this, we also drop the note
about the special situation on SSH in the introduction. This is already
mentioned in the SSH section.

Reported-by: Bogdan Prădatu <bogdan.pradatu@gmail.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>