Set a deterministic date when committing patches
Using the current date on committing patches will lead to different
commit hashes on every build and will invalidate the bitbake
recipe cache.

Try to extract the author date from the patch file. If this is not
possible, because a raw diff is used, fall back to the filesystem
timestamp.

Signed-off-by: Tobias Schaffner <tobias.schaffner@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>