kas-isar: Install reprepro from proposed-updates
The reprepro tool in Debian has a critical bug, making it impossible to
work on zstd compressed .deb packages. As this is a host tool, and
Ubuntu uses zstd compression as default, this bug breaks building Ubuntu
targets in the kas-container.

A patched version of the tool is already available in
bookworm-proposed-updates and will be added in the next point release.
Until then, we take it from proposed-updates. Once upstream releases
Debian 12.8, this patch should be dropped again.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>