refactor detection of CI systems
kas has support to behave differently on well-known CI systems (e.g.
automatically inject tokens and configure git). To add support for
further managed environments, we carve this detection logic out and add
a test for it. While doing so, we also replace the string matching by
using an enum.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>