ci: make pinning of buildkit-cache-dance more strict
The buildkit-cache-dance action is not an official action. By that, we
should pin it to an exact sha, as we do with other external actions.
While changing this, we also update to 3.1.2, which fixes issues around
wrong uid/gids of the cache artifacts.

Changelog: https://github.com/reproducible-containers/buildkit-cache-dance/releases/tag/v3.1.2

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>