docs: document kas-container command
The kas-container script is the last undocumented tool of the kas suite.
To change this, we add a page in the html documentation and a manpage.

As the kas-container script is written in bash, the sphinx-build
automodule plugin cannot be used to extract the parameters. To work
around this, we run the kas-container script during the docs build,
extract the usage information and massage that a bit to be RST compliant
(and look nice). Then, the generated files are included in the
documentation templates.

As the kas-container docs are generated on the fly, we also must ignore
missing-files errors (D000 Problems with "include" directive path) when
statically checking the doc tree with doc8.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>