deprecate non full-length digests for git repos
The documentation was not precise enough w.r.t. what exactly is a commit
id. By that, it was possible to use abbreviated git commit ids. This
creates problems down the line, as these might not be unique and also
comparisons between full-length and abbreviated ones are not trivial.

As kas users might have used this, we deprecate non full-length commit
ids and will remove support for them in some future version.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>