attestation: make repo digest more precise
Previously we always denoted the digest as sha-1 in the attestation
file. However, this only holds for git repos and even there might not be
correct anymore as git supports sha-256 commit ids. To fix this, we
we use the digest type "gitCommit" (for git repos), which supports both
full sha-1 and sha-256 digests. For mercurial, no standard type is
defined, so we use the custom type "mercurialCommit" (which is allowed by
the format).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>