repos: cache properties that run external commands
Some of our properties internally run external VCS commands. As this can
be quite costly, we now cache the result. This is correct, as the repo
properties are referential transparent within a single kas invocation.

As cached_property is not available for python<3.8, we add a dependency
to the retrofit implementation in cached-property.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>