repos: order layers by repo and layer name
Usually the order in which layers are appended to the bblayers does not
matter, but there are cases where fine-grained control is needed.

Before this patch, all layers were sorted by the final path. This
however creates issues when specifying both the path and the name, as
well as when absolute paths are used.

We change this by first sorting by repo name and then layer path. We
further document this behavior in the project configuration.
As the repo name was already part of the full path before, it is very
unlikely that downstream users are affected (the order does not change
in most cases).

Fixes: #130

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>