tests: split online and offline tests
When packaging kas it is important to not run the online tests, as the
test environment (or distro standards) prohibit any external resources.
In the past this was tricky to achieve, leading to (almost) completely
disabled test suites (e.g. in Debian).

To fix this, we mark all tests that require internet access with the
"online" marker. In the GitHub actions CI these are also executed
seperately to ensure tests are marked correctly. While doing so, we
document this split and also state that offline tests are preferred.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>