fix: dump repos by key instead of name
In most cases the repo name is identical to the repo id. However, this
only holds if no explicit name is set. The dump plugin reconstructed the
flattened repo dict by name, which is wrong as it is not identical to
the input.

This is fixed by getting the repos with their keys (instead of just the
repos) and properly reconstructing the dict based on the keys.

Fixes: #131

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>