tests: clone bitbake from GitHub mirror
This mirror is much faster, especially when cloning from within the
GitHub Actions CI. By switching over using KAS_PREMIRRORS, we also test
the premirrors logic of kas. The example is also quite close to the real
world use case of replacing the official OE repos with the faster
mirrors on GitHub.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>