cleanup logging environment after kas invocation
When running the tests, kas.kas is invoked multiple times.
As this function takes care of setting up the logging environment, we
also register the log handler multiple times. This results in duplicated
log output (n times) in the pytest log and console capture. Also, it
results in a nested exception on test error, completely spamming the
console:

  ValueError: I/O operation on closed file.

To fix this, we explicitly cleanup the logging environment after leaving
the kas.kas function. By that, we can no longer log in the entrypoint
(the function that calls kas.kas), so we handle the exceptions in
kas.kas itself, print the log messages and re-throw into the calling
context to terminate with the corresponding error code.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>