do not catch SIGINT under pytest
When running under pytest, we must not register a signal handler for
SIGINT (CTRL-C), as otherwise the pytest internal signal handling does
not work. This was also the root-cause, why we could not properly
terminate our tests by pressing CTRL-C.

Without that handler, the SIGINT raises a KeyboardInterrupt, which is
handeled by pytest and either stops the test execution or prints a
precise stacktrace of the current execution (which is valuable for
debugging hanging tests).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>