replace use of kas.run_cmd in tests
The kas.run_cmd function internally uses the async variant that requires
an event loop. As a preparation to get rid of the ambient event loop
(fix for Python 3.12), we need to port all external users over
subprocess.

As run_cmd is used quite often in test_refspec we just provide a
replacement implementation. By that we do not disturb the history too
much.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>