run explicit event loop to fix issues on Python 3.12
When running the tests under Python 3.12, a deprecation warning is
issued around non-proper handling of the event loop. Currently, we
always assume there is an ambient event loop. However, the asyncio
implementation just spawns one in case there is none yet. This behavior
is deprecated in Python 3.12 and likely will be an error in the future.

To fix this, we explicitly create our event loop. By that, we take care
of the lifetime of the loop and need to close it before returning from
the kas.kas function.

During normal execution, by latest the atexit handler takes care of the
termination, but during tests pytest emits exceptions in-between two kas
runs in the same process. By that, the exceptions are raised in the
event loop, bubble up to pytest and are caught there. Thus we left the
kas.kas function (not closing the loop) and on the next kas invocation
a new loop is created overwriting the old one. In the destruction of
the old one the remaining exceptions are thrown, but this time in a
non-raisable context which is reported as an error. To fix this, we
wrap the command processing in a try-finally block to make sure the
loop is always properly terminated.

Once Python 3.6 support is dropped, the whole event-processing should
be ported over to async tasks, which are easier to handle, especially
in the exception / cancellation case.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>