GitLab CI: make ssh config check less strict
For backwards compatibility we only inject the ssh rewrites if no ssh
config is present. The documentation was a bit imprecise w.r.t. what
exactly ssh config means, but the implementation checked for any file in
~/.ssh. This is too strict, as the authorized_keys or known_hosts files
are sometimes used to inject the ssh fingerprint of the own GitLab
instance. The presence of these files does not interfere with our
rewrite rules.

We change that by only checking for kas related ssh config vars as well
as the presence of `~/.ssh/config`.

Reported-by: Florian Bezdeka <florian.bezdeka@siemens.com>
Fixes: af6b9ae3 ("auto-inject git credentials on gitlab ci")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Reviewed-by: Frieder Schrempf <frieder.schrempf@kontron.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>