consider git submodules as part of main repo
All kas include files on the cmdline need to come from the same
repository. If git submodules are used, this check made it impossible to
add config snippets on the cmdline that are stored in the submodule.
However, the version of the sub-module is pinned by the gitmodules file
of the main project, so it can be considered safe to include these.

We now change that logic to always consider the super (topmost parent)
repository to be the root repo. This is a slight behavior change in case
the kas repo itself was checked out as a submodule. However, this
use-case was anyways never officially supported.

Reported-by: Urs Ritzmann <ursritzmann@protonmail.ch>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>