move lock file handling to new lock plugin
The helpers to create and update kas project config lockfiles have been
part of the dump plugin. As it is planned to add more logic around
locking, this will lead to a clumsy to use command line of that plugin.

We now carve out the inplace update logic (the option --inplace) to a
new plugin "lock". For backwards compatibility, the dump --lock --inplace
combination is re-routed to the new plugin and a deprecation warning is
shown. The documentation and the kas-container is updated as well.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>