add override to remove default tag or branch
Currently there is no way to remove a default branch / tag value for a
single repository. This makes it impossible to add a repository by
referencing a commit that is not on a branch (hence the branch key
cannot be set).

We fix this by also allowing the nulltype value null as valid values for
both branch and tag. This change is fully backward compatible as the
value was not allowed before and also does not change semantics.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>