fix resource leaks on command execution error
The Macro runner executes a list of setup commands, user commands and
teardown commands. In the previous implementation, these commands were
just combined and executed one-after-another. Once the first command
failed (raised an exception), the whole execution was terminated without
running the cleanup commands. This led to resource leaks of external
processes (like the kas-spawned ssh-agent).

We fix this by pairing setup and cleanup commands. Whenever a setup
command is executed successfully, it's corresponding cleanup command is
preprended to the cleanup list. By that, the cleanup commands are
executed in the reverse order of the setup. In case an exception is
raised during setup or user command execution, the corresponding cleanup
commands are executed in the finally handler.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>