make version property in kas json schema more strict
Previously all integers and the string "0.10" were supported. We make
this more strict by limiting the schema to versions which are actually
supported (e.g. no later versions than x). This helps users that use an
IDE with schema support to learn about the latest supported schema.

We further load the supported schema versions from the schema itself to
have a single source of truth.

Proposed-by: Konrad Weihmann <konrad.weihmann@avnet.eu>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>