enhance json schema with descriptions
When writing kas config files in an IDE that supports json schemas (like
vscode), the IDE gives hints about possible keys / values. We improve
this by adding short descriptions to the nodes. As the draft-04 version
does not yet support the "deprecated" property, we just mention
deprecations in the description.

No functional change.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>