fix(GitRepo): use git diff-index to check if dirty
git status -s fails in VSCode devcontainers, when users have
status.branch = true set in their global .gitconfig, as the branch name
is always printed and the dirty-check checks for any stdout output.

This would not be a problem, if we could check the return code instead,
but hg status will exit with zero. Otherwise, we could add the --quiet
option to the git is_dirty_cmd.

Signed-off-by: Oliver Kästner <okaestner@rosen-nxt.com>