use max header version number when merging configs
The header version cannot be merged in-order, as otherwise an include
file with a lower version defines the version of the merged config,
which might not be valid (e.g. if properties are used that are not
available in that version).

This aspect was irrelevant prior to introducing the dump plugin, as the
data is not used within kas. However, the dump plugin needs to use the
max header version as otherwise the generated config might not be valid.

Fixes: 5b3e23812 ("add plugin to dump flattened config")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>