fix(GitRepo): use git diff --stat to check if dirty
This check previously was changed to use git diff-index to check if a
working tree contains tracked-but-changed files. However, the diff-index
command only compares the results against the git index but does not
re-scan the filesystem (the index is eventually consistent). By that,
the check incorrectly reported a repo as dirty if the index was not
updated yet.

We now switch to the high-level command "git diff" which takes care of
updating the index beforehand. This command also is not affected by the
"status.branch = true" config option, which was the reason we switched
to git diff-index.

Fixes: 0694dd670c5 ("fix(GitRepo): use git diff-index to check if dirty")

Reported-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>