schema: make validation error message less verbose
In case the config does not validate against the schema a very long
error message was printed which even included the schema definition
itself (per single error). This has proven to be not very helpful,
as it made the actual error hard to spot.

We change this by following the upstream recommendation of sorting the
errors and also only printing the error message (which only contains the
affected node), reducing the error message from >300 to a couple
of lines. When running in debug mode, we explicitly dump the schema on
error, but only once for all errors.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>