schema: print link to online documentation on validation error
As the schema itself as a documentation on which config elements are
supported has proven to not be very helpful, we now directly print a
link to the online documentation of the project configuration on error.

While this somehow links kas to its online representation, this pattern
became common practice for many OSS projects.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>