docs: document rules for provided directories
Having overlapping kas dirs (via the env vars KAS_<x>_DIR) creates
trouble regarding file permissions and cleaning. This was never
really supported, but it was also not explicitly forbidden. We now
state that all passed directories must be non-overlapping, except if
stated otherwise. Using sub-dirs of the KAS_WORK_DIR is allowed, as
this is a common use-case.

We further state which directories must exist, which parent dirs must
exist and which are created recursively.

No functional change.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>