schema: replace use of oneOf by anyOf
The schema definition "oneOf" is discouraged if not explicitly needed,
as it requires the checker to check against all variants to ensure that
only exactly one matches. It also makes it harder for IDEs to provide
meaningful auto-completions. As our schema is written in a way that all
matches are exclusive anyways, we can safely switch to anyOf.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>