ci: run next workflow on all branches except master
The GitHub Actions CI is a valuable tool to find errors which are hard
to detect locally (e.g. errors on old Python versions or in
kas-container). Previously, the CI was only run on the next branch,
which made it cumbersome to test multiple series in parallel.

We now execute parts of the next workflow on all branches (except
master). The steps that modify the global state (like the container
upload or build attesting) are skipped if not on next.

Proposed-by: Michael Adler <michael.adler@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>