menu: make kconfiglib weak dependency
In our docs we state that kconfiglib is an optional dependency. However,
it actually never was optional as we include the kconfiglib
unconditionally.

As the kconfiglib dependency is only needed for the menu plugin, make it
optional. By that, it is easier to derive container images for CI
systems where the menu plugin is not needed.

As kconfiglib is no longer a mandatory dependency, it also will not
automatically get installed in CI. For that, we add a new test target
that collects all test dependencies and call that in the CI.

Fixes: 4f23045ce ("Add kconfiglib-based menu plugin")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Tested-by: Michael Adler <michael.adler@siemens.com>