run includehandler tests with kas context
The includehandler tests directly invoke kas internals, but these
transitively call functions (like run_cmd) that require a context.
In case the tests are run directly, without running other kas
functions upfront that create a context, they fail. This can be
reproduced with: pytest -v -k includehandler

We now add a autouse fixture that creates the context upfront and
destroys it after the test was executed.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>