kas-container: properly handle workdir-relative mounts
Previously the kas directories that can be controlled by setting an
env-var were always mounted to dedicated mountpoints inside the
container. This however creates problems if these are overlapping on the
host as they look isolated from inside the container while they are
not.

We fix this by checking if a provided path is relative to the work
directory (which is always mounted rw) and just adjust the environment
variables inside the container accordingly, without using a dedicated
mountpoint. We further align the (recursive) creation of the passed
directories with kas (and bitbake in case of DL_DIR and SSTATE_DIR).

While changing this, we also replace all invocations of readlink with
the corresponding realpath variant as this simplifies the review.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>