kas-container: run clean commands inside container
As kas is now able to perform the various clean commands and also the
special handling outside of the container is getting too complex on
rootless container execution, we just forward the cleaning task to kas
inside the container.

By that, all clean-related special handling in the kas-container is
removed as well.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>